// src/app/page/master/map-angka-to-hari/map-angka-to-hari.interface.ts
import { EventEmitter } from '@angular/core';

export interface MapAngkaToHari {
	kdProduk?: string;
	kdEvent?: number;
	label?: any;
	value?: any;
	kdMapping?: any;
	kode?: any;
	kdProfile?: string;
	version?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
}